use an ArrayHashMap for tracking active variables
Instead of an ArrayList of sets which we search in linear order, we can
instead use a single ArrayHashMap, which stores its entries in insertion
order. We track the position where each frame began inserting entries
into the map so we can truncate the array once the frame is popped.